requirements.txt
environs==14.1.1
requests==2.32.3
python-telegram-bot==13.15